fix campaign 1-2-3 integration

+++ src/controller/order_controller.go
@@ -12,8 +12,9 @@ import (
 )
 
 var (
-	ordered int
+	ordered_count int 
 	totPrice float64
+	order []entities.Order
 )
 
 func ListOrders() gin.HandlerFunc {
@@ -25,13 +26,8 @@ func ListOrders() gin.HandlerFunc {
 		if str_order == nil{
 			c.JSON(http.StatusOK," NO ORDERED ")
 		}else {
-			// restrict dublicate order by get list
-			if items == nil{
-				ordered++
-			}
 			if str_order != nil{
 				str_order := session.Get("orders").(string)
-				var order []entities.Order
 				err := json.Unmarshal([]byte(str_order), &order)
 				if err != nil{
 					log.Println(err,"Unmars ListOrder")
@@ -40,7 +36,7 @@ func ListOrders() gin.HandlerFunc {
 				data := map[string]interface{}{
 					"Order": order,
 					"TotalOrderPrice" : totPrice,
-					"TotalOrder" : ordered,
+					"TotalOrder" : ordered_count,
 				}
 				c.JSON(http.StatusOK, data)
 			}
@@ -50,38 +46,45 @@ func ListOrders() gin.HandlerFunc {
 
 func BuyCart() gin.HandlerFunc {
 	return func(c *gin.Context) {
-		
-		session := sessions.Default(c)
-		
-		var totQua int64
-		//var totprice float64
-		var order []entities.Order
 
-		if campaign == 0. {
-			totPrice += totVat
-		}else {
-			totPrice = campaign
-		}
+		if items != nil {
 
-		order = append(order, entities.Order{Order_ID: 1, Orderered_At: entities.Date{Day: time.Now().Day(),Mount: time.Now().Month(), Year: time.Now().Year()}, TotalAmount: 100000,TotalAmountinMonth: 3000 , Quantity: totQua})
-		bytesOrder, err := json.Marshal(order)
-		if err != nil{
-			log.Println(err)
-		}
-		session.Set("orders", string(bytesOrder))
+			ordered_count++
+			session := sessions.Default(c)
+			
+			var totQua int64
+			//var totprice float64
+			var order []entities.Order
 
-		err3 := session.Save()
-		if err3 != nil{log.Println(err3)}
-		
-		for _, id := range items{
-			deleteErr := CART.DeleteItem(id.Id)
-			tot = 0
-			totVat = 0
-			if err != nil{
-				log.Println(deleteErr,"DeleteAll")
+			if campaign == 0. {
+				totPrice += totVat
+			}else {
+				totPrice += campaign
 			}
+
+			order = append(order, entities.Order{Order_ID: 1, Orderered_At: entities.Date{Day: time.Now().Day(),Mount: time.Now().Month(), Year: time.Now().Year()}, TotalAmount: 100000,TotalAmountinMonth: 3000 , Quantity: totQua})
+			bytesOrder, err := json.Marshal(order)
+			if err != nil{
+				log.Println(err)
+			}
+			session.Set("orders", string(bytesOrder))
+
+			err3 := session.Save()
+			if err3 != nil{log.Println(err3)}
+			
+			for _, id := range items{
+				deleteErr := CART.DeleteItem(id.Id)
+				tot = 0
+				totVat = 0
+				campaign = 0
+				if err != nil{
+					log.Println(deleteErr,"DeleteAll")
+				}
+			}
+			items = cartModel.ListAll()
+			c.Redirect(http.StatusFound, "/users/orders")
+		}else {
+			c.JSON(http.StatusBadRequest,"YOUR CART IS EMTY")	
 		}
-		
-		c.Redirect(http.StatusFound, "/users/orders")
 	}
 }
